Move GetDefaultTsapPair to TsapPair class
